fix: updates input schemas
