+++ serverless.yml
@@ -1,4 +1,4 @@
-service: juanCastroArancibia-practicoIntegrador
+service: juanArancibia-practicoIntegrador
 
 provider:
   name: aws
@@ -17,6 +17,7 @@ functions:
       environment:
         CLIENTS_TABLE: !Ref ClientsTable
       iamRoleStatementsInherit: true
+      iamRoleStatementsName: juanCastroArancibia-integrador-assign-gift
       iamRoleStatements: 
         - Effect: Allow
           Action:
@@ -32,6 +33,7 @@ functions:
       name: ${self:service}-${self:provider.stage}-assign-card-function
       environment:
         CLIENTS_TABLE: !Ref ClientsTable
+      iamRoleStatementsName: juanCastroArancibia-integrador-assign-card
       iamRoleStatementsInherit: true
       iamRoleStatements: 
         - Effect: Allow
