+++ src/create-client/create-client.yml
@@ -5,6 +5,7 @@ CreateClient:
   environment:
     CLIENTS_TABLE: !Ref ClientsTable
     CLIENTS_CREATED_TOPIC: !Ref ClientCreatedTopic
+  iamRoleStatementsName: juanCastroArancibia-integrador-create-client
   iamRoleStatementsInherit: true
   iamRoleStatements: 
     - Effect: Allow
