+++ src/create-client/domain/createClient.js
@@ -1,12 +1,25 @@
-const { ValidateCreateClientInput } = require('../schema/createClientInput')
+const { ValidateCreateClientInput } = require('../schema/input/createClient.input')
+const { ClientCreated } = require('../schema/event/clientCreated.event')
 const { createClient } = require('../service/createClient')
 const { publishNewClient } = require('../service/publishClientCreated')
 
 module.exports = async (commandPayload, commandMeta) => {
   new ValidateCreateClientInput(commandPayload, commandMeta)
 
+  if (
+    calculateAge(commandPayload.birth) < 18 ||
+    calculateAge(commandPayload.birth) > 65
+  ) {
+    return {
+      statusCode: 400,
+      body: "Client must be between 18 and 65 years old",
+    };
+  }
+
   await createClient(commandPayload)
-  await publishNewClient(commandPayload)
+
+  const clientCreatedEvent = new ClientCreated(commandPayload)
+  await publishNewClient(clientCreatedEvent)
 
   return {
     status: 200,
